updTE

+++ utils/mutation.server.ts
@@ -2,7 +2,6 @@ import { MutationResolvers } from './types.generated.server'
 import { prisma } from 'utils/prisma.server'
 import bcrypt from 'bcrypt'
 import { User } from '@prisma/client'
-import { authenticated } from './auth.server'
 
 const Mutation: MutationResolvers = {
   async register(_parent, args, ctx) {
@@ -46,7 +45,9 @@ const Mutation: MutationResolvers = {
     return null
   },
 
-  createListItem: authenticated(async (_parent, args, ctx) => {
+  async createListItem(_parent, args, ctx) {
+    if (!ctx.req.session.user?.id) return null
+
     const { listItemInput } = args
     const { bookId, ...rest } = listItemInput
 
@@ -54,7 +55,7 @@ const Mutation: MutationResolvers = {
       data: {
         ...rest,
         User: {
-          connect: { id: ctx?.req?.session?.user?.id },
+          connect: { id: ctx.req.session.user.id },
         },
         Book: {
           connect: { id: bookId },
@@ -63,9 +64,11 @@ const Mutation: MutationResolvers = {
     })
 
     return 'list item created'
-  }),
+  },
 
-  updateListItem: authenticated(async (_parent, args, ctx) => {
+  async updateListItem(_parent, args, ctx) {
+    if (!ctx.req.session.user?.id) return null
+
     const { listItemInput } = args
     const { id, bookId, ...rest } = listItemInput
 
@@ -74,9 +77,11 @@ const Mutation: MutationResolvers = {
       data: { ...rest },
     })
     return 'update list item'
-  }),
+  },
 
-  removeListItem: authenticated(async (_parent, args, ctx) => {
+  async removeListItem(_parent, args, ctx) {
+    if (!ctx.req.session.user?.id) return null
+
     const { id } = args
 
     const remove = await prisma.listItem.delete({
@@ -86,7 +91,7 @@ const Mutation: MutationResolvers = {
     })
 
     return 'list item removed'
-  }),
+  },
 }
 
 export default Mutation
